Rename QuizOverview and Quizresultat components to QuizOverviewView and QuizResultView
